SecureFileCrypt
SecureFileCrypt is a lightweight encryption and decryption tool designed for securing your sensitive files. With an intuitive interface, it allows you to easily encrypt files for secure storage or sharing and decrypt them when needed. The app employs strong encryption algorithms to ensure the security and privacy of your data. It also includes features like password configuration and customizable file location settings for added convenience and control. SecureFileCrypt is ideal for individuals and businesses looking for a reliable solution to protect their confidential information.